VMimage pre-release tests fix
It removes downloading of all vmimage versions. It will try to download
only the default version to check that downloading is working.

Reference: #5619
